edit.bump-revision: Try to keep comment blocks about the variable intact
PORTREVISION is normally deleted and readded in a cleanup
and reordering attempt since some ports have it in an
unexpected location. I think that this is still a good thing
to do. However it goes wrong when PORTREVISION is preceded
by a comment block. It is better to just reuse the old
PORTREVISION location in such cases.

Fixing it in edit.bump-revision is easier than in edit.merge
since there are fewer corner cases to consider.