Add support for a WWW macro
There have been multiple proposals to move the URL of the project
website from the last line of pkg-descr into the ports' Makefiles:

	https://reviews.freebsd.org/D2949
	https://reviews.freebsd.org/D34979

I have got permission from PortMgr to implement this change, but
before committing this change to bsd.port.mk and the port Makefiles,
this change should be supported by portfmt and portlint.

The placement of WWW= between the BLOCK_MAINTAINER and BLOCK_LICENSE
has been agreed to by PortMgr (in a mail from tcberner).

In fact, I plan to create the patch that adds WWW= to all Makefiles
using portedit.